Update the docker build.
Run `npm install` from the `pg/htdocs` location.  Also both of the `npm install` commands need to be run with `--unsafe-perm` so that the `prepare` script will run.
Don't try to change ownership and permissions on the now non-existent applets directory.
Fix the bin/addcourse script for the new WWSafe.pm location.
Add the Ubuntu package libyaml-libyaml-perl which is needed for the new PGEnvironment.pm module.