Add the table-responsive wrapper to the table on the add users page.
Fix the comparison of the request parameter that determines if the
"Import from where?" select on the homework sets editor page is multiple
or not on page load.  That needs to be a string comparison.  Also,
change that to checking that the value is `ne '1'` instead of `eq '8'`
in case that in the future the multiple size is changed.  The single
size probably won't.

Make the state of the checkboxes and set id select on `Scoring Tools`
page (Scoring.pm) persistent when the form is submitted.  Also make it
so that the "You must select one or more sets for scoring!" message is
only displayed after clicking on the "Score selected set(s) and save
to:" button, and not when the page initially loads.

Fix a typo in Achievements.pm.  aria_labe -> aria_label.  Also prefix
the id's of the hidden authen inputs in the forms for using achievement
items inside the modal dialogs that are embedded in the page when a user
has such items available for use.  Really, these hidden forms shouldn't
even have id's.  However, at this point some javascript uses that to
find these forms.  That can be fixed by searching by name instead, but
will take care to make sure all case where these id's are used are
found.

Remove the duplicate id "problem_body" in the system.template file.

Remove the `role="navigation"` attribute from the breadcrumb `nav` in
the system.template and gateway.template files.  `nav` tags should not
have that role.

Remove the `<meta http-equiv="X-UA-Compatible" content="IE=Edge"/>` tag
from the system.template and gateway.template files.  Those tags are
neither valid nor needed anymore.

Use the `$setID` with underscores for the html id of the set siblings on
the ProblemSet page.  The `$pretty_set_id` has spaces which is invalid
for an html id.  Do these anchors even need an id?

Move the progress bar out of the siblings `ul` in Problem.pm.  It is
invalid html to have that in the list.  Also correct the html markup for
the divs that make up the progress bar.  Several were self closing, and
then also followed by an end div tag.

Make the `problem-sub-header` `div` in the page title of Problem.pm a
`span` instead (but give it the css style display block).  It is invalid
html for a `div` to be inside an `h1`.

Remove the underscore from the target attribute for the show me another
link in Problem.pm.  Targets should not have the underscore prefix
unless they are reserved word targets like `_self` or `_blank`.  Also
change that target name to be more consistent with the other target
names used by webwork.

Fix the number of columns for the gateway quiz rows that have no
versions, or that do not allow the display of scores for the set in
Grades.pm.