Update the javascript for SetMaker (the library browser).
This fixes several bugs.

One issue is one that I caused when I switched from using the
`onclick` attributes and inline javascript to using `addEventListener`
and external javascript.  I missed a case where the "More/Less like
this" handling changed the `onclick` attribute.  That has now been
converted to change the appropriate data attribute instead.

Another issue was that the "Target Set" select was directly using its
values for labels.  In particular it was using the "Select a Set from
this Course" value which is translated.  The javascript looks for that
value, and if it is translated it fails.  So now proper labels and
values are used.  Furthermore the labels for sets have the underscore
translated to spaces.  In addition, all of the "All ..." options for the
subjects, chapters, sections, etc., are now translated in labels.  Of
course the values are not translated.

In problemsetdetail.js there is a minor issue fixed in that the executor
of a promise should not be an async function.

Translations are also added for some of the dialogs on the library
browser page (the ones that will typically appear in day-to-day use).
These are passed to the javascript via data attributes.