Add the capability of deleting a specific quiz version.
This has been requested in the forums (see for instance
https://webwork.maa.org/moodle/mod/forum/discuss.php?d=2400), and is
something that I would also find useful.  In my classes I have quizzes
that allow three versions with one submission per version, and show only
one problem per page.  Students frequently will start one version, and
then accidentally hit "Grade Test" after only working one problem.  This
wastes one of the versions for this student.  Prior to this pull request
I could do this, but had to do so manually by deleting the version and
problems in the database.  Note that doing so also does not cause
problems anywhere else.  Note that prior to pull request #1585 there was
a minor glitch caused on the user detail page if a version was deleted
that was not the last version that exists.  (This frequently happens
because not only do students accidentally hit submit, but then they go
start another version before contacting me to fix the problem.)  The
glitch was caused by the assumption that the version_id stored in the
database coincided with array indices (offset by one) in the array of
versioned user sets obtained by listing all versions in the database.
That of course is not the case with #1585.  I have been deleting
versions manually for years with no issues other than this one.

This adds a checkbox to the quiz versions on the user detail page (which
you can get to by clicking on the assigned sets for a user in the class
list editor).  If this checkbox is unchecked, then that specific version
of the gateway quiz will be deleted for that user.  Note that javascript
is used to make the server side behavior clear to the user.  If the
template set is unchecked, then all of the versions are unchecked as
in addition to the template set being unassigned, the versions will all
be deleted.  If a version is checked (after previously being unchecked)
then it is ensured that the template set is also checked.

Since this adds a javascript file for the UserDetail.pm module, all of
the inline javascript in that file is also moved into the javascript
file (the change, blur, and keyup handlers for the date inputs, and the
click handler for the "Assign All Sets to Current User" button).