If a specific version of a gateway quiz is requested, then use its open date rather than the template set's open date to determine if it is open.
This allows for things like having a gateway quiz that is open for a
period of time, then closed for a period, and then subsequetly open
again.  With this pull request, students that had taken the quiz during
the first period that the quiz was open can still view their past
attempts during the period in the middle that the quiz is closed.

This was requested by @glarose for a situation that came up for him.