Fix the tag widget.
I broke this when I removed underscore from the vendor js files.  I
forgot about this one completely unnecessary usage of that javascript
library.  That usage has been removed.  Furthermore, this removes all of
the related inline javascript.  The contents of the
tagging-taxonomy.json file are now fetched by the javascript, so this
also removes an instance of disk access in the perl code.

Also remove some javascript and css that is being loaded in
ProblemSetDetail.pm and SetMaker.pm that isn't needed with the current
rendering methods.