Add a new permission `change_pg_display_settings`.
If a user does not have this permission, then the user will not be able
to change the display settings on the `User Settings` page.

I don't want my students changing these settings.  There really is no
need for them to do so.  Also, changing the MathQuill setting from `Yes`
to `No` can cause problems.  Basically, if the student has that set to
`No`, and the professor has it set to `Yes`, then when the instructor
acts as that student and views the problems for that student, no answers
appear in the MathQuill answer inputs.  The answers are in the now
hidden text inputs only.  Perhaps it should be made so that when an
instructor acts as a student, the student's user settings for display
are used.

Furthermore, change the display on the `User Settings` page.  If a user
does not have a permission to change a setting, don't show "You do not
have the permission to change this setting."  Instead just don't show
the setting or its header at all.  There are three permissions for this:
`change_password`, `change_email_address`, and the new
`change_pg_display_settings`.

Additionally, if a user doesn't have any of those permissions, don't
even add the `User Settings` link in the site nav (MAIN MENU).