Fix show me another for problems with resources added via PGresource.
The problem is that any resource (like an image) added in this way adds
the url with unique id alias (from PGalias.pm) of the resource to the
generated html (the pg body_text).  This will always be different if the
seed is different.  So the check in ShowMeAnother.pm will always
determine that the body_text of the new problem and the original problem
are different regardless of if there are any actual differences
elsewhere.

So this pull request strips those urls from both the original problem
and the new problem before comparison.

Also add the showMeAnotherBox style back.  That was accidentally removed
in the bootstrap update pull request.