Add an asset build system for javascript and css files.
All javascript and scss/css files in htdocs/themes and htdocs/js/apps are compiled to minimized js or cs files.  The filename of the generated file contains a hash of the contents.  The generated file names are saved to the file htdocs/static-assets.json identified by the unminimized file name.  Webwork then loads this file (and a same named file in pg/htdocs), and any time that a javascript or css file is needed a new method `getAssetsURL` in lib/WeBWorK/Utils.pm is called to get the minimized file name from these lists.  Note the template files call the `url` method of ContentGenerator.pm, which in turn calls `getAssetsURL`.  The file name is now an argument to the `url` method, instead of being appended to the return value of this method.

Most important here is that since the generated file names contain the content hash, this solves the browser caching problem.  Any time the contents of a javascript or css file change, the hash will change.  Since the file name is now different the browser will not look in its cache for the updated file, and will download the new one.

Note that files in webwork2/htdocs and pg/htdocs are directed to their respective actual server urls (/webwork2_files or /pg_files respectively).  So the need for the apache2 rewrite rule is lessened.  It is still left in place as a fallback.

In addition, there is an added file `htdocs/third-party-assets.json` which contains a mapping of files we use in `htdocs/node_modules` (installed via `npm install`) to a corresponding CDN location (on cdn.jsdelivr.net).

When `npm install` is run, the `prepare` script is run, which in turn executes the `generate-assets` script, and that runs the node script `htdocs/generate-assets.js`.  One can also separately run that script with `./generate-assets.js` or `node generate-assets.js`.  If the option `-useCDN` is passed to that script the CDN locations will be used instead of the local files for third party libraries.  Note that the old `generate-css` and `css-prefix` npm scripts have been removed.  Thos will no longer work with this scheme, and the `generate-assets` script does everything they did and more.

Note that all references to the math4-coloring.css file have been removed.  That file is now obsolete.  The _theme-colors.scss file should be used instead.  In addition, the `generate-assets.js` script will create a math4-overrides.css and math4-overrides.js file in each theme directory if one does no exist already.  So this solves the console warnings about those files not being found when the html2xml rendering method is used.  The partial fix for this in the docker-entrypoint.sh file has been removed.  That will be more properly dealt with when `npm install` is run in the docker build.

The `.min` infix is used on the generated js and css files so that these can be ignored in `.gitignore`.