Add `for` attributes to labels and matching ids to corresponding inputs on the course admin page.
Add labels to the model courses that are the options for the "Copy
templates from" select in the add course form.  The empty string value
has the translated label 'No Course'.

Remove a stray end table tag in the rename course form.

Add a missing end form tag in the edit location form.

Remove the obsolete `tt` tags, and use the bootstrap utility class
`font-monospace` instead.