Add `for` attributes to labels and matching ids to inputs in ScrollingRecordList.pm.  Also add an id to the main scrolling list of the record, and make the label be the header of the scrolling record list.
Add `for` attributes to labels and matching ids to inputs in
Assigner.pm, SendMail.pm, and ShowAnswers.pm.

Remove the gray color style on the time stamp for answers in the past
answers table in ShowAnswers.pm.  That has color contrast issues even
against the white background.

Make the dropped user color darker in UserList.pm.  Also make the user
id the label for the checkbox for selecting the user.  The select all
checkbox needs a label and it also creates an unlabeld column header.
Not sure what to do about this.  For now the checkboxes are switched to
being created with CGI::input of type checkbox (rather than the
shorthand CGI::checkbox) so that these can be created without an empty
label which is invalid html.

Do the same thing for the checkboxes in ProblemSetList.pm,
AchievementList.pm, and UsersAssignedToSet.pm.

Also in AchievementList.pm make the achievement id the second column and
the enabled column the third.  That order did not make sense.

Make the invalid `example` attribute on the new set input in SetMaker.pm
a placeholder instead.  Also fix an invalid value of 1 for disabled
inputs.

Make sure that the student progress page always has a title.  The
template forces the existence of the `h1` and it is invalid for it not
to have content.  Also replace the invalid `align` attribute on the
headers of the sets and students with the bootstrap css class
`text-center`.  This is all done on the statistics page as well.

On the course config page (Config.pm) make the setting description the
label for each setting (except for checkbox lists), and uri escape the
variable for the help buttons to deal with hash key config variables
properly.  Also make the active tab in the nav a `span` instead of an
`a`.  This is probably a false positive, but it makes the firefox
accessibility tool not complain.

Add the `table-responsive` div wrapper for the table on the
AchievementList.pm editor page.  Also remove the deprecated linebreaks
attribute used, and replace the vspace attribute with css.

Replace more tables that should not be tables with modern markup and
bootstrap styling on the ProblemSetDetail.pm page.

Remove some invalid width attributes on inputs.

Also remove the obsolete `font` tags in AchievementList.pm.  In general
AchievementList.pm is a disaster area and there is a lot to be done
there.

Fix a persistent typo in the action tabs.  Its aria-labelledby, not
aria-labelled-by (after underscore is translated).

Remove obsolete `tt` tags, and use the bootstrap utility class
`font-monospace` instead.