Improve CodeMirror with some new options.
Add an option to turn spell checking on or off, an option to select the
CodeMirror theme, and an option to change the keymap.

These options are saved in local storage, and so are persistent as long
as you are using the same browser and device.

Also, the limitation on the width of the codemirror div has been
increased.  The size will now be as large as it can up to 1000px.

Note that the code that was duplicated for the Achievement Editor and
the PG Problem Editor is now moved into a separate file that is called
by both of those modules.