If a course has sections or recitations,  then add a dropdown to the right of the previous dropdown in the student nav that appears when an instructor is acting as another user and is viewing a problem in a problem set or viewing a gateway quiz version.
This dropdown lists all sections and recitations for users assigned to
the set on the problem page and for taken test version on a gateway quiz
page.  If the section or recitation is clicked on, then only those
users/tests will be shown.  There is also an item added in the dropdown
when a filter is active that if clicked will remove the filter.  The
active filter will be persistent when navigating between problems and
students, or when a problem or gateway quiz submit button is pressed.

This is a feature that was requested in #1240.

In addition the order of the things in the sticky nav is changed on the
problem page.  Now the "Previous Problem", "Problem List", and "Next
Problem" buttons are on the left (or top on narrow screens), and the
student nav is on the right (or bottom on narrow screens).  This makes
the "Previous Problem", "Problem List", and "Next Problem" buttons
appear in the same place that they were before the student nav was
added.