Revamp the PGProblemEditor
The problem editor now renders the problem (or header file) immediately
when the page loads in an iframe directly on the page, instead of the
hack of loading a complete webwork page in an iframe in a modal dialog.
That led to all of the crud of having to hide the main menu and other
page elements that looked odd embedded in a dialog.

When the "View/Reload" tab is selected and "Take Action" is clicked and
"Open in new window" is not checked the problem is reloaded with content
updated from the CodeMirror editor's current contents.  If a submit
button is pressed in the problem the contents are also updated with the
CodeMirror editor's current contents.  Also, when "Randomize Seed" is
clicked the problem reloads immediately with the new seed.  In any of
these cases a temporary file is saved.  A new WebworkWebservice method
has been added for this purpose when rendering in the same window.  The
javascript uses this to save the temporary file.  The method only allows
saving in a course's templates directory and only if the user has
the permission to modify_problem_sets.

Note that the "View/Reload" tab is now shown for course info files.  I
don't know why that was disabled, it was fully functional.

When editing hardcopy header files, the in page renderer just gives a
message that hardcopy header contents can only be viewed in a new
window, and the "Open in new window" checkbox is checked by default.

If a pg file is saved as a new version, then auxiliary files are also
copied, even if the file basename and directory name are not the same.
That really didn't make sense to only copy the auxiliary files in that
case.  Usually that will not be the case anyway, and the resulting
copied file will fail to work without its resources.

The completely unused options_info file editing capability has been
removed.

Add a sanity check to the PG problem editor to check that the file that
the user is saving matches the file on record.  This adds a hidden input
to the save form that stores the file name that was in the database for
the problem when the editor is opened.  When the user attempts to save
the file, that path is verified.  If the path does not match what is in
the database at that time, then refuse to save the file.  This fixes
issue #1580.

The "Report Bugs in this Problem" button is now only shown for problems
in the OPL or Contrib.

Convert WebworkClient formats to Mojolicious templates.  Now there are
only three template files, default.html.ep, default.json.ep, and
ptx.html.ep.  The simple, sticky, and default output formats all use the
default.html.ep template.  The ptx output format clearly uses the
ptx.html.ep template.  The json format uses the default.json.ep
template.  The raw format is still just code in formatRenderedProblem.
The useless body_text output format has been removed, as well as the
dysfunctional tex output format.  Note the FormatRenderedProblem module
no longer uses CGI at all.

Two new options for submitting the problem source are added to the
renderProblem method.  The first is raw unnencoded source (the parameter
is rawProblemSource). This will not work embedded in html as is often
done with the formats for embedding in an iframe, but can be done when
form submission is handled by javascript.  The second is uri encoded
source (the parameter is uriEncodedProblemSource). One could Use
encodeURI in javascript to encode the source.  Server side that is
decoded using Mojo::Util url_unescape.  This does work for embedding the
source in the html.  Base64 encoding really should never be used, and
just needs to go away.

Finally, a new action tab has been added to the editor page.  A
"Generate Hardcopy" tab.  This tab is visible when editing a problem or
set header.  There are options to select the seed, hardcopy format (pdf
or text), and harcopy theme (the themes set in conf files).  Note that
if the pdf format is selected and there are errors then a zip file will
be offered for download, otherwise of course a pdf file will be offered.
If the tex format is selected then a zip file will be sent.  Another new
WebworkWebservice option is added for this.  This replaces the
dysfunctional tex output format.  It is called hardcopyRenderedProblem
(not sold on that name).  It works much like formatRenderedProblem,
except it generates a pdf or zip file.  First renderProblem is called
and then, if the outputformat is tex or pdf hardcopyRenderedProblem is
called.  If a problem results in no body text, there are translator
errors, or in the extreme event that neither a pdf or zip can be
created, then a text error is rendered instead.

Note that a more reliable method for adding resource files to the zip
archive for the tex format is used and is also implemented in
Hardcopy.pm.