When hardcopy files are generated, stop generating them in a publicly accessible location.
This solves the problem of files being made publicly available that may contain secure information.

The way this works is that the files are generated in a temporary subdirectory of /opt/webwork/webwork2/tmp/courseId/hardcopy/userID. Of course since that is not a publicly accessible, this means that the files can also not be directly downloaded via a link.  Instead the files are served via the content generator method `reply_with_file`.

When hardcopy generation succeeds the file is sent immediately, and is deleted from the server.  Note that file may be a pdf or zip file depending on the hardcopy format chosen.

If hardcopy generation fails, then the temporary directory is not deleted (same as before), but the files are served by adding a url parameter to the hardcopy module path whose value is the file name.  The hardcopy module then serves these files via `reply_with_file`.

Note that users will only be able to download files that are in the subdirectory with their own userID.  Furthermore users that do not have the 'download_hardcopy_multiuser' permission will not be able to download any of these files unless the user id that they are signed in with is in the file name between periods.  For example if `tguy` is signed in, then the user will be able to download `courseId.tguy.setId.pdf`, but not the temporary files `work.aq8jglRJ/pdflatex.stderr`, `work.aq8jglRJ/hardcopy.tex`, etc. (the file must also be in the subdirectory named for the active user id).

Note that when acting as another user the files are still created in a subdirectory named for your user id, and you can still download them.  So you technically do not have access to files generated by your students via the ui.  Of course you can still act as a student and get essentially the same files.

Also note that what your browser does when these files are served will probably change depending on how you have your browser set.  If pdf files opened in the browser before, they may not now (although this can still be configured by your browser if it has such configuration).  Text files will probably not open in your browser and will need to be downloaded.