Fix numerous display and html validation issues on the instructor stats page.
Generally clean Stats.pm up.  That file was a considerable mess.  Many
variables were defined and never used, and generally a lot of code was
unneeded.

Note that the borders on the svg image that are now appearing were in
the code before, but were not appearing because of an error in the style
attribute.  That was fixed, and the styles were also converted to proper
svg attributes.

For gateway quizzes, the links to problems now work.  Note that there is
no valid link to the actual problem in the gateway quiz.  So instead the
Problem.pm editMode is used.  This at least gives the instructor a link
that works and shows the problem which is really what is important here.