Add the role "figure" to the progress bar divs in lib/WeBWorK/ContentGenrator/Problem.pm and lib/WeBWorK/ContentGenrator/Achievements.pm.  It is invalid for a standard html element like a div to have an aria label without such a role.  I am not sure that this is the correct role for this though.
Fix a `div` that should have been a `start_div` in lib/WeBWorK/ContentGenerator/ProbemSet.pm.

Remove the obsolete summary attributes on the tables in lib/WeBWorK/ContentGenerator/ProbemSet.pm and lib/WeBWorK/ContentGenerator/ProbemSets.pm.  Those tables are structured in such a way that the summaries are not needed.  The summaries were basically just saying what is already in the captions and headers.  The summary in ProblemSets.pm was also mentioning the buttons below the table which was not correct to do to begin with.

The obsolete summary attributes in lib/WeBWorK/ContentGenrator/Instructor/UserList.pm and lib/WeBWorK/ContentGenrator/Instructor/ProblemSetList.pm have also been removed.  Some of their contents were added to the "Site Description" instead.

Make the "Login Name" column header contents the label for the select all checkbox in lib/WeBWorK/ContentGenrator/Instructor/UserList.pm, lib/WeBWorK/ContentGenrator/Instructor/ProblemSetList.pm, and lib/WeBWorK/ContentGenrator/Instructor/AchievementList.pm.  Also add an aria label to the select all checkbox itself which serves as the text for the column header (the WAVE tool does not accept this, but the Firefox accessibility tool and Chrome Lighthouse do, and it is better than nothing).

Fix an `i` tag that was not properly closed in lib/WeBWorK/ContentGenrator/Instructor/AchievementList.pm.

In lib/WeBWorK/ContentGenrator/Instructor/UserList.pm and lib/WeBWorK/ContentGenrator/Instructor/ProblemSetList.pm make the column headers the labels for the inputs in the column using the aria_labelledby attribute.  Of course this means that ids had to be added to the columnn headers.

Fix potentially duplicated ids in the achievement item forms inside the modal dialogs in lib/WeBWorK/ContentGenrator/Achievements.pm.  These will be duplicated if a student has more than one item available.

Remove the config strings from lib/WeBWorK/Localize.pm, and directly mark them in conf/defaults.config.  The bin/update_localization_files script is updated to pick up on these.  Having the strings duplicated in Localize.pm was not a good idea.

Also remove the translations for site descriptions (and removed table summaries) from the lexicon and put them directly into the one place that each is used.  This is better for translators if they want to look in source files for context.  The lexicon is meant to be used for translations that are used in multiple locations and need to be kept in sync.

The config values for the LTI tab that were in conf/authen_LTI.conf.dist have been moved to conf/LTIConfigValues.config.  Instead the @LTIConfigVariables has been added to conf/authen_LTI.conf.dist.  If any of the values in that array are uncommented, then the LTI tab will be shown with the uncommented values appearing.  The conf/LTIConfigValues.config file is included in conf/authen_LTI.conf after the definition of @LTIConfigVariables so that it can act on that setting.